Initial commit
Forgot to initial commit at the beginning but i am now!

I've set up the folder structure, and began working on the front end, the landing page is mainly done, with header/footer components and a hero section and hero badge row done.

 Not quite complete but im pretty happy with it so far.

I am logging off for the night and sharing my progress with the group